i hope this works

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -9,7 +9,7 @@
 
 size_t print_list(list_t *h)
 {
-	int node_count = 0;
+	size_t node_count = 0;
 
 	while (h)
 	{
